backend/czi_hosted/requirements.txt
anndata>=0.7.0
boto3>=1.12.18
click>=7.1.2
fastobo>=0.6.1
Flask>=1.0.2
Flask-Compress>=1.4.0
Flask-Cors>=3.0.6
Flask-RESTful>=0.3.6
flask-server-timing>=0.1.2
flask-talisman>=0.7.0
flatbuffers>=1.11.0
flatten-dict>=0.2.0
fsspec>=0.4.4,<0.8.0
gunicorn>=20.0.4
h5py<3.0.0 # h5py>=3.0.0 had a breaking change; there is a fix in anndata>=0.7.5
numba>=0.49.1,<0.53.0
numpy>=1.15.0
packaging>=20.0
pandas>=1.0,!=1.1  # pandas 1.1 breaks tests, https://github.com/pandas-dev/pandas/issues/35446
PyYAML>=5.3
scipy>=1.0
requests>=2.22.0
tiledb>=0.5.9,>=0.6.2,!=0.7.2
s3fs==0.4.2
scanpy==1.4.6 # Until we move to anndata 0.7.4 scanpy needs to be pinned here
sqlalchemy>=1.3.18
umap-learn<0.5.0 # The pinned version scanpy is not compatible with latest umap-learn
